// frontend/src/shared/api/generated/models/contract-details-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Pied Piper Freelance API
 * API для биржи фриланса (MVP)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface ContractDetailsDTO {
    'contractId'?: number;
    'jobId'?: number;
    'jobTitle'?: string;
    'freelancerId'?: number;
    'freelancerName'?: string;
    'clientId'?: number;
    'clientName'?: string;
    'totalAmount'?: number;
    'status'?: string;
    'cancellationRequestedBy'?: number;
    'cancellationRequesterName'?: string;
    'cancellationRequestedAt'?: string;
    'cancellationReason'?: string;
    'createdAt'?: string;
}